+++ src/pages/Parkmoonyoung/Main/MainPmy.js
@@ -1,12 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import './MainPmy.scss';
 import { Link } from 'react-router-dom';
-import RepleLi from './RepleLi';
+import FeedsArea from './FeedsArea';
 import { Data_Info_Area } from './Data_Info_Area';
 
 export default function MainPmy() {
-  const [isFeedsData, setFeedsData] = useState([]);
-
   useEffect(() => {
     fetch('/data/feedsData.json')
       .then(response => {
@@ -18,28 +16,7 @@ export default function MainPmy() {
       });
   }, []);
 
-  /*******   ******/
-  const [isRepleValue, setRepleValue] = useState('');
-
-  const [repleData, setRepleData] = useState([
-    { name: 'myp', text: '거봐 내가 좋다고 했지', index: 0 },
-  ]);
-
-  const renderReple = repleData.map(function (data, index) {
-    return <RepleLi data={data} key={index} />;
-  });
-
-  // 인풋 밸류 체크
-  function valueCheck(e) {
-    setRepleValue(e.target.value);
-  }
-
-  // 게시 버튼 클릭시 실행되는 함수
-  function addPost() {
-    setRepleValue('');
-
-    setRepleData([...repleData, { name: 'myp', text: isRepleValue }]);
-  }
+  const [isFeedsData, setFeedsData] = useState([]);
 
   return (
     <div className="wrap">
@@ -100,80 +77,7 @@ export default function MainPmy() {
           <div className="feedsLoopArea">
             {/* 피드 반복 구간 */}
             {isFeedsData.map(function (feeds) {
-              return (
-                <div className="feedsArea" key={feeds.id}>
-                  <div className="feedsHeader">
-                    <div className="profileBox">
-                      <div className="imgBox">
-                        <img src={feeds.userThumbImgUrl} alt="myp" />
-                      </div>
-                      <span className="nickName">{feeds.userId}</span>
-                    </div>
-                    <button className="btn detail">
-                      <span className="blind">상세보기</span>
-                    </button>
-                  </div>
-
-                  <div className="feedsContentsArea">
-                    <div className="feedsImgArea">
-                      <ul>
-                        <li>
-                          <img src={feeds.userFeedsImgUrl} alt="myp" />
-                        </li>
-                      </ul>
-                    </div>
-                    <div className="feedsUtilArea">
-                      <div className="feedsBtnArea">
-                        <button className="btn heart active">
-                          <span className="blind">좋아요</span>
-                        </button>{' '}
-                        <button className="btn reple">
-                          <span className="blind">리플쓰기</span>
-                        </button>
-                        <button className="btn share">
-                          <span className="blind">공유하기</span>
-                        </button>
-                        <button className="btn favorite">
-                          <span className="blind">즐겨찾기</span>
-                        </button>
-                      </div>
-                      <div className="repleTotal">
-                        <div className="imgBox">
-                          <img
-                            src="/images/Parkmoonyoung/profile/my-profile.jpg"
-                            alt="myp"
-                          />
-                        </div>
-                        <p className="totalTxt">
-                          <span className="firstlyNick">aineworld</span>님{' '}
-                          <span className="totalNumber">외 10명</span>이
-                          좋아합니다
-                        </p>
-                      </div>
-                      <div className="myReple reduce">
-                        <b className="myName">myp</b>
-                        <span className="txt">
-                          위워크에서 진행한 베이킹 클래스...
-                        </span>
-                        <button className="btn more">더 보기</button>
-                      </div>
-                      <ul className="repleList">{renderReple}</ul>
-                    </div>
-                    <div className="addRepleBox">
-                      <input
-                        type="text"
-                        className="replePost"
-                        placeholder="댓글 달기..."
-                        onChange={valueCheck}
-                        value={isRepleValue}
-                      />
-                      <button className="btn post" onClick={addPost}>
-                        게시
-                      </button>
-                    </div>
-                  </div>
-                </div>
-              );
+              return <FeedsArea data={feeds} key={feeds.id} />;
             })}
           </div>
